rag-level-evaluation/requirements.txt
# Database connectivity
psycopg[binary]>=3.2.2

# OpenAI API
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0

# Machine learning (for ROC-AUC calculation)
scikit-learn>=1.0.0

# Progress bars
tqdm>=4.66.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
numpy>=1.24.0
